third_party/WebKit/LayoutTests/media/track/track-cue-gc-wrapper-expected.txt

Tests that added cue object wrappers live across garbage collections.

** Add a text track to the video element **
EXPECTED (track.hasOwnProperty('custom') == 'true') OK
** Add cues with own native property to the track with enter event listener. **
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK

** Trigger a garbage collection. **
EXPECTED (track.hasOwnProperty('custom') == 'true') OK

** Play the video and test cue wrappers. **
RUN(video.play())
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
EXPECTED (cue.hasOwnProperty('custom') == 'true') OK
END OF TEST
